accountbooks/requirements.txt
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
distro==1.6.0
Django==3.2.10
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
jsonschema==3.2.0
mysqlclient==2.1.0
paramiko==2.9.2
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.8
websocket-client==0.59.0
